# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pratilipi_Microservices",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "localhost://"
    }
  ],
...
  "paths": {
    "/000/content/upload": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "content/upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "content": [
                    {
                      "title": "Harry Potter",
                      "Totalchapters": 10,
                      "chapters": [
                        {
                          "chNo": 1,
                          "chName": "episode1",
                          "content": "aaaaaaaaaaaaaa"
                        },
                        {
                          "chNo": 2,
                          "chName": "episode2",
                          "content": "bbbbbbbbbbbb"
                        },
                        {
                          "chNo": 3,
                          "chName": "episode3",
                          "content": "cccccccccccc"
                        },
                        {
                          "chNo": 4,
                          "chName": "episode4",
                          "content": "dddddddddddd"
                        }
                      ]
                    },
                    {
                      "title": "spider-man",
                      "Totalchapters": 10,
                      "chapters": [
                        {
                          "chNo": 1,
                          "chName": "episode1",
                          "content": "aaaaaaaaaaaaaa"
                        },
                        {
                          "chNo": 2,
                          "chName": "episode2",
                          "content": "bbbbbbbbbbbb"
                        },
                        {
                          "chNo": 3,
                          "chName": "episode3",
                          "content": "cccccccccccc"
                        },
                        {
                          "chNo": 4,
                          "chName": "episode4",
                          "content": "dddddddddddd"
                        }
                      ]
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/000/content/usercontent": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "content/usercontent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "userId": "624a07980ab6a78f15f99922"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  }
}